[cmake] Provide core library as object library
This is an effort to make the core library more suited to cmake
when it comes to dealing with the cyclic dependency between core
and platform layers. By using the object library, the platform lib
can decide to include the core objects within it. This prevents
having to double link the platform and core libs, which acts in
a tempremental way due to the there being only one variation of
the plat, but 2 variations of the core. The application only then
has to link to the platform ot library, and can ignore the core.